restructured folders, added comp,serv,and tests to modules folders
