--- src/app/components/profile/profile.component.html
@@ -1,1 +1,0 @@
-<p>profile works!</p>
