--- src/app/components/profile/profile.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-profile',
-  templateUrl: './profile.component.html',
-  styleUrls: ['./profile.component.css'],
-})
-export class ProfileComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit(): void {}
-}
