--- src/app/components/submission-form/submission-form.component.html
@@ -1,1 +1,0 @@
-<p>submission-form works!</p>
