--- src/app/components/submission-form/submission-form.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-submission-form',
-  templateUrl: './submission-form.component.html',
-  styleUrls: ['./submission-form.component.css'],
-})
-export class SubmissionFormComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit(): void {}
-}
