--- src/app/components/submission-history-vehicle/submission-history-vehicle.component.html
@@ -1,1 +1,0 @@
-<p>submission-history-vehicle works!</p>
