--- src/app/components/submission-history/submission-history.component.html
@@ -1,1 +1,0 @@
-<p>submission-history works!</p>
