--- src/app/components/submission-history/submission-history.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-submission-history',
-  templateUrl: './submission-history.component.html',
-  styleUrls: ['./submission-history.component.css'],
-})
-export class SubmissionHistoryComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit(): void {}
-}
