+++ src/app/modules/inventory/components/vehicle-list/vehicle-list.component.html
