+++ src/app/modules/inventory/components/vehicle/vehicle.component.html
