+++ src/app/modules/inventory/inventory.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { VehicleListComponent } from 'src/app/components/vehicle-list/vehicle-list.component';
-import { VehicleComponent } from 'src/app/components/vehicle/vehicle.component';
+import { VehicleListComponent } from 'src/app/modules/inventory/components/vehicle-list/vehicle-list.component';
+import { VehicleComponent } from 'src/app/modules/inventory/components/vehicle/vehicle.component';
 
 @NgModule({
   declarations: [VehicleListComponent, VehicleComponent],
