+++ src/app/modules/inventory/services/InventoryAbstract.ts
@@ -1,7 +1,7 @@
 import { Observable, of } from 'rxjs';
 // @ts-ignore
-import vehicles from '../tests/data/vehicles.json';
-import Vehicle, { convertVehicle } from '../models/vehicle';
+import vehicles from '../../../tests/data/vehicles.json';
+import Vehicle, { convertVehicle } from '../../../models/vehicle';
 
 export abstract class InventoryServiceAbstract {
   public vehicles: Vehicle[] = vehicles.map((data) =>
