+++ src/app/modules/profile/components/profile/profile.component.html
@@ -1,0 +1,1 @@
+<p>profile works!</p>
