+++ src/app/modules/profile/components/submission-form/submission-form.component.html
@@ -1,0 +1,1 @@
+<p>submission-form works!</p>
