+++ src/app/modules/profile/components/submission-form/submission-form.component.ts
@@ -1,0 +1,12 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-submission-form',
+  templateUrl: './submission-form.component.html',
+  styleUrls: ['./submission-form.component.css'],
+})
+export class SubmissionFormComponent implements OnInit {
+  constructor() {}
+
+  ngOnInit(): void {}
+}
