+++ src/app/modules/profile/components/submission-history-vehicle/submission-history-vehicle.component.html
@@ -1,0 +1,1 @@
+<p>submission-history-vehicle works!</p>
