+++ src/app/modules/profile/components/submission-history-vehicle/submission-history-vehicle.component.ts
@@ -1,0 +1,12 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-submission-history-vehicle',
+  templateUrl: './submission-history-vehicle.component.html',
+  styleUrls: ['./submission-history-vehicle.component.css'],
+})
+export class SubmissionHistoryVehicleComponent implements OnInit {
+  constructor() {}
+
+  ngOnInit(): void {}
+}
