+++ src/app/modules/profile/components/submission-history/submission-history.component.html
@@ -1,0 +1,1 @@
+<p>submission-history works!</p>
