+++ src/app/modules/profile/components/submission-history/submission-history.component.ts
@@ -1,0 +1,12 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-submission-history',
+  templateUrl: './submission-history.component.html',
+  styleUrls: ['./submission-history.component.css'],
+})
+export class SubmissionHistoryComponent implements OnInit {
+  constructor() {}
+
+  ngOnInit(): void {}
+}
