+++ src/app/modules/profile/services/vehicle-submission.service.ts
@@ -1,0 +1,8 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root',
+})
+export class VehicleSubmissionService {
+  constructor() {}
+}
