--- src/app/services/vehicle-submission.service.ts
@@ -1,8 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root',
-})
-export class VehicleSubmissionService {
-  constructor() {}
-}
